requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
gunicorn==21.2.0
pytest==8.0.2
black==24.1.1
flake8==7.0.0
alembic==1.13.1
SQLAlchemy-Utils==0.41.1 